Fix: getRootPropsFix
