+++ src/app/dao/components/createProposalModal.tsx
@@ -1,39 +1,22 @@
 
 'use client';
-import { MarkdownRenderers } from "@/app/upload/utils/MarkdownRenderers";
-import { uploadFileToIPFS } from "@/app/upload/utils/uploadToIPFS";
 import {
     Box,
     Button,
-    Center,
     HStack,
-    Input,
-    Spinner,
-    Tooltip,
     VStack, useBreakpointValue
 } from "@chakra-ui/react";
 import { Web3Provider } from '@ethersproject/providers';
-import snapshot from '@snapshot-labs/snapshot.js';
-import MDEditor, { commands } from '@uiw/react-md-editor';
 import { useEffect, useMemo, useState } from "react";
-import { useDropzone } from "react-dropzone";
-import { FaImage, FaSave } from "react-icons/fa";
-import ReactMarkdown from 'react-markdown';
-import rehypeRaw from 'rehype-raw';
-import rehypeSanitize from 'rehype-sanitize';
-import remarkGfm from 'remark-gfm';
 import CreateProposalConfirmationModal from "./createProposalConfirmationModal";
-
-
+import ProposalEditor from "./ProposalEditor";
+import ProposalPreview from "./ProposalPreview";
 
 const hub = 'https://hub.snapshot.org';
-const client = new snapshot.Client712(hub);
 
 interface CreateProposalModalProps {
     connectedUserAddress: string;
 }
-
-
 
 const CreateProposalModal = ({ connectedUserAddress }: CreateProposalModalProps) => {
     const [spaceInfo, setSpaceInfo] = useState({});
@@ -45,58 +28,6 @@ const CreateProposalModal = ({ connectedUserAddress }: CreateProposalModalProps)
     const PINATA_GATEWAY_TOKEN = process.env.NEXT_PUBLIC_PINATA_GATEWAY_TOKEN;
     const [isConfirmationModalOpen, setIsConfirmationModalOpen] = useState(false);
 
-    const { getRootProps, getInputProps } = useDropzone({
-        noClick: true,
-        noKeyboard: true,
-        onDrop: async (acceptedFiles) => {
-            setIsUploading(true);
-            for (const file of acceptedFiles) {
-                const ipfsData = await uploadFileToIPFS(file); 
-                if (ipfsData !== undefined) { 
-                    const ipfsUrl = `https://ipfs.skatehive.app/ipfs/${ipfsData.IpfsHash}?pinataGatewayToken=${PINATA_GATEWAY_TOKEN}`;
-                    const markdownLink = file.type.startsWith("video/") ? `<iframe src="${ipfsUrl}" allowfullscreen></iframe>` : `![Image](${ipfsUrl})`;
-                    setValue(prevMarkdown => `${prevMarkdown}\n${markdownLink}\n`);
-                }
-            }
-            setIsUploading(false);
-        },
-        accept: {
-            'image/*': ['.png', '.gif', '.jpeg', '.jpg'],
-            'video/*': ['.mp4', '.mov']
-        },
-        multiple: false
-    }
-    );
-    const extraCommands = [
-        {
-            name: 'uploadImage',
-            keyCommand: 'uploadImage',
-            buttonProps: { 'aria-label': 'Upload image' },
-            icon: (<Tooltip label="Upload Image or Video"><span><FaImage color="yellow" /></span></Tooltip>),
-            execute: (state: any, api: any) => {
-               
-                const element = document.getElementById('md-image-upload');
-                if (element) {
-                    element.click();
-                }
-            }
-        },
-        {
-            name: 'saveDraftInTxt', 
-            keyCommand: 'saveDraftInTxt', 
-            buttonProps: { 'aria-label': 'Save Draft' },
-            icon: (<Tooltip label="Save Draft" ><span><FaSave color="limegreen" /></span></Tooltip>),
-            execute: (state: any, api: any) => {
-                const element = document.createElement('a');
-                const file = new Blob([value], { type: 'text/plain' });
-                element.href = URL.createObjectURL(file);
-                element.download = "draft.txt";
-                document.body.appendChild(element);
-                element.click();
-            }
-        }
-
-    ];
     const fetchCurrentBlockNumber = async () => {
         const blockNumber = await web3.getBlockNumber();
         setCurrentBlockNumber(blockNumber);
@@ -125,95 +56,36 @@ const CreateProposalModal = ({ connectedUserAddress }: CreateProposalModalProps)
     return (
         <Box>
             {useBreakpointValue({
-                base: <VStack spacing="4">
-                    <Box
-                        width="100%" 
-                    >
-                        <Input
-                            placeholder="Proposal Title"
-                            value={title}
-                            onChange={(e) => setTitle(e.target.value)}
-                            borderColor={"green.600"}
-                            border={"2px solid"}
-                            color={"limegreen"}
-                            _placeholder={{ color: "limegreen", opacity: 0.4 }}
-                            focusBorderColor="limegreen"
+                base: (
+                    <VStack spacing="4">
+                        <ProposalEditor
+                            value={value}
+                            setValue={setValue}
+                            title={title}
+                            setTitle={setTitle}
+                            setIsUploading={setIsUploading}
+                            PINATA_GATEWAY_TOKEN={PINATA_GATEWAY_TOKEN}
                         />
-                        <Box  {...getRootProps()} >
-                            {isUploading && <Center><Spinner /></Center>}
-
-                            <MDEditor
+                        <ProposalPreview value={value} />
+                    </VStack>
+                ),
+                md: (
+                    <HStack>
+                        <Box width="50%">
+                            <ProposalEditor
                                 value={value}
-                                onChange={(value: any) => setValue(value || "")}
-                                commands={[
-                                    commands.bold, commands.italic, commands.strikethrough, commands.hr, commands.code, commands.table, commands.link, commands.quote, commands.unorderedListCommand, commands.orderedListCommand, commands.codeBlock, commands.fullscreen
-                                ]}
-                                extraCommands={extraCommands}
-                                previewOptions={{ rehypePlugins: [[rehypeSanitize]] }}
-                                height="600px"
-                                preview="edit"
-                                style={{
-                                    border: "2px solid limegreen",
-                                    padding: "10px",
-                                    backgroundColor: "black",
-                                }}
+                                setValue={setValue}
+                                title={title}
+                                setTitle={setTitle}
+                                setIsUploading={setIsUploading}
+                                PINATA_GATEWAY_TOKEN={PINATA_GATEWAY_TOKEN}
                             />
                         </Box>
-                    </Box>
-                    <Box
-                        width="100%"
-                        height="100%"
-                        border="1px solid limegreen"
-                        minHeight="100%"
-                    >
-                        <ReactMarkdown
-                            components={MarkdownRenderers}
-                            remarkPlugins={[remarkGfm]}
-                            rehypePlugins={[rehypeRaw]}
-                        >
-                            {value}
-                        </ReactMarkdown>
-                    </Box>
-                </VStack>, md: <HStack>
-                    <Box
-                        width="50%"
-                    >
-                        <Box marginTop="3" {...getRootProps()} >
-                            {isUploading && <Center><Spinner /></Center>}
-
-                            <MDEditor
-                                value={value}
-                                onChange={(value: any) => setValue(value || "")}
-                                commands={[
-                                    commands.bold, commands.italic, commands.strikethrough, commands.hr, commands.code, commands.table, commands.link, commands.quote, commands.unorderedListCommand, commands.orderedListCommand, commands.codeBlock, commands.fullscreen
-                                ]}
-                                extraCommands={extraCommands}
-                                previewOptions={{ rehypePlugins: [[rehypeSanitize]] }}
-                                height="600px"
-                                preview="edit"
-                                style={{
-                                    border: "1px solid limegreen",
-                                    padding: "10px",
-                                    backgroundColor: "black",
-                                }}
-                            />
+                        <Box width="50%">
+                            <ProposalPreview value={value} />
                         </Box>
-                    </Box>
-                    <Box
-                        width="50%"
-                        height="100%"
-                        border="1px solid limegreen"
-                        minHeight="100%"
-                    >
-                        <ReactMarkdown
-                            components={MarkdownRenderers}
-                            remarkPlugins={[remarkGfm]}
-                            rehypePlugins={[rehypeRaw]}
-                        >
-                            {value}
-                        </ReactMarkdown>
-                    </Box>
-                </HStack>
+                    </HStack>
+                )
             })}
             <Button
                 width={"100%"}
@@ -237,8 +109,15 @@ const CreateProposalModal = ({ connectedUserAddress }: CreateProposalModalProps)
             >
                 Log Space Details
             </Button>
-            {isConfirmationModalOpen && <CreateProposalConfirmationModal proposalBody={value} isOpen={isConfirmationModalOpen} connectedUserAddress={connectedUserAddress} onClose={() => setIsConfirmationModalOpen(false)} title={title} />}
-
+            {isConfirmationModalOpen && (
+                <CreateProposalConfirmationModal
+                    proposalBody={value}
+                    isOpen={isConfirmationModalOpen}
+                    connectedUserAddress={connectedUserAddress}
+                    onClose={() => setIsConfirmationModalOpen(false)}
+                    title={title}
+                />
+            )}
         </Box>
     );
 };
